+++ src/packages/badge/badge.harmony.css
@@ -7,7 +7,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  background: linear-gradient(135deg, #FF0F23 0%, #FF0F23 100%);
+  background: linear-gradient(135deg, #ff475d 0%, #ff0f23 100%);
   top: -10%;
   right: -15%;
   padding: 3px;
@@ -27,7 +27,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  background: linear-gradient(135deg, #FF0F23 0%, #FF0F23 100%);
+  background: linear-gradient(135deg, #ff475d 0%, #ff0f23 100%);
   color: #fff;
   padding: 0 4px;
   font-size: 10px;
@@ -53,6 +53,11 @@
   border: 0px solid #ffffff;
   border-radius: 7px;
   padding: 0;
+}
+.nut-badge-outline .nut-badge-content {
+  background: #ffffff;
+  border: 1px solid #ff0f23;
+  color: #ff0f23;
 }
 
 [dir=rtl] .nut-badge-icon,
