+++ src/packages/switch/demo.tsx
@@ -6,6 +6,7 @@ import Demo3 from './demos/h5/demo3'
 import Demo4 from './demos/h5/demo4'
 import Demo5 from './demos/h5/demo5'
 import Demo6 from './demos/h5/demo6'
+import Demo7 from './demos/h5/demo7'
 
 const SwitchDemo = () => {
   const [translated] = useTranslate({
@@ -15,6 +16,7 @@ const SwitchDemo = () => {
       asyncControl: '受控',
       customColor: '自定义颜色',
       supportText: '支持文字',
+      supportIcon: '支持Icon',
       eventTip: '触发了 onChange 事件，开关状态：',
     },
     'zh-TW': {
@@ -23,6 +25,7 @@ const SwitchDemo = () => {
       asyncControl: '受控',
       customColor: '自定義顏色',
       supportText: '支持文字',
+      supportIcon: '支持Icon',
       eventTip: '觸發了 onChange 事件，開關狀態：',
     },
     'en-US': {
@@ -31,6 +34,7 @@ const SwitchDemo = () => {
       asyncControl: 'controlled',
       customColor: 'Custom Color',
       supportText: 'Support Text',
+      supportIcon: 'Support Icon',
       eventTip: 'Emit onChange event, current state:',
     },
   })
@@ -43,12 +47,14 @@ const SwitchDemo = () => {
         <Demo2 />
         <h2>{translated.disabled}</h2>
         <Demo3 />
-        <h2>{translated.eventTip}</h2>
-        <Demo4 />
-        <h2>{translated.customColor}</h2>
-        <Demo5 />
         <h2>{translated.supportText}</h2>
+        <Demo4 />
+        <h2>{translated.supportIcon}</h2>
+        <Demo5 />
+        <h2>{translated.eventTip}</h2>
         <Demo6 />
+        <h2>{translated.customColor}</h2>
+        <Demo7 />
       </div>
     </>
   )
