+++ src/packages/switch/demos/h5/demo4.tsx
@@ -1,16 +1,15 @@
 import React from 'react'
-import { Cell, Switch, Toast } from '@nutui/nutui-react'
+import { Cell, Switch, Space } from '@nutui/nutui-react'
 
 const Demo4 = () => {
-  const onChange = (value: boolean, event: any) => {
-    Toast.show(`触发了onChange事件，开关状态：${value}`)
-  }
   return (
     <Cell>
-      <Switch
-        defaultChecked
-        onChange={(value, event) => onChange(value, event)}
-      />
+      <Space>
+        <Switch defaultChecked activeText="开启" inactiveText="关闭" />
+        <Switch defaultChecked activeText="开" inactiveText="关" />
+        <Switch defaultChecked activeText="开启" inactiveText="关闭" disabled />
+        <Switch activeText="开启" inactiveText="关闭" disabled />
+      </Space>
     </Cell>
   )
 }
