+++ src/packages/switch/demos/h5/demo5.tsx
@@ -1,16 +1,11 @@
 import React from 'react'
 import { Cell, Switch } from '@nutui/nutui-react'
+import { Check, Close } from '@nutui/icons-react'
 
 const Demo5 = () => {
   return (
     <Cell>
-      <Switch
-        defaultChecked
-        style={{
-          '--nutui-switch-open-background-color': 'blue',
-          '--nutui-switch-close-line-background-color': '#ebebeb',
-        }}
-      />
+      <Switch defaultChecked activeText={<Check />} inactiveText={<Close />} />
     </Cell>
   )
 }
