+++ src/packages/switch/demos/h5/demo6.tsx
@@ -1,10 +1,16 @@
 import React from 'react'
-import { Cell, Switch } from '@nutui/nutui-react'
+import { Cell, Switch, Toast } from '@nutui/nutui-react'
 
 const Demo6 = () => {
+  const onChange = (value: boolean, event: any) => {
+    Toast.show(`触发了onChange事件，开关状态：${value}`)
+  }
   return (
     <Cell>
-      <Switch defaultChecked activeText="开" inactiveText="关" />
+      <Switch
+        defaultChecked
+        onChange={(value, event) => onChange(value, event)}
+      />
     </Cell>
   )
 }
