+++ src/packages/switch/demos/taro/demo6.tsx
@@ -1,10 +1,20 @@
 import React from 'react'
 import { Cell, Switch } from '@nutui/nutui-react-taro'
+import Taro from '@tarojs/taro'
 
 const Demo6 = () => {
+  const onChange = (
+    value: boolean,
+    event: React.MouseEvent<Element, MouseEvent>
+  ) => {
+    Taro.showToast({ title: `触发了onChange事件，开关状态：${value}` })
+  }
   return (
     <Cell>
-      <Switch defaultChecked activeText="开" inactiveText="关" />
+      <Switch
+        defaultChecked
+        onChange={(value, event) => onChange(value, event)}
+      />
     </Cell>
   )
 }
