+++ src/packages/switch/switch.harmony.css
@@ -1,39 +1,40 @@
 .nut-switch {
-  cursor: pointer;
+  position: relative;
   display: flex;
   flex-direction: row;
   align-items: center;
-  width: 40px;
-  height: 24px;
-  line-height: 24px;
+  min-width: 46px;
+  height: 28px;
+  line-height: 28px;
   background-color: #ff0f23;
-  border-radius: 8px;
+  border-radius: 50px;
   background-size: 100% 100%;
   background-repeat: no-repeat;
   background-position: center center;
   flex: 0 0 auto;
 }
 .nut-switch-button {
+  position: absolute;
+  top: 2px;
   display: flex;
   flex-direction: row;
   align-items: center;
   justify-content: center;
-  height: 20px;
-  width: 20px;
-  transform: translateX(2px);
-  border-radius: 6px;
+  height: calc(28px - 2px * 2);
+  width: calc(28px - 2px * 2);
+  border-radius: 50%;
   background: #ffffff;
-  transition: transform 0.3s;
-  box-shadow: 0 0 transparent;
+  transition: left 0.3s;
+  box-shadow: 0px 2px 6px 0px rgba(0, 0, 0, 0.4);
 }
 .nut-switch-button-open {
-  transform: translateX(18px);
+  left: calc(100% - 28px + 2px);
 }
 .nut-switch-button-open-rtl {
   transform: translateX(-18px);
 }
 .nut-switch-button-close {
-  transform: translateX(2px);
+  left: 2px;
 }
 .nut-switch-button-close-rtl {
   transform: translateX(-2px);
@@ -42,30 +43,37 @@
   background-color: #c2c4cc;
 }
 .nut-switch-close-line {
-  width: 12px;
+  width: calc((28px - 2px * 2) / 2);
   height: 2px;
-  background: #fff;
+  background: #ffffff;
   border-radius: 2px;
+}
+.nut-switch-label {
+  display: inline-flex;
+  align-items: center;
+  height: 100%;
+  white-space: nowrap;
+  color: #ffffff;
+  font-size: 12px;
+}
+.nut-switch-label-open {
+  margin: 0 calc(28px - 2px + 5px) 0 7px;
+}
+.nut-switch-label-open-rtl {
+  transform: translateX(18px);
+}
+.nut-switch-label-close {
+  margin: 0 7px 0 calc(28px - 2px + 5px);
+}
+.nut-switch-label-close-rtl {
+  transform: translateX(-18px);
+}
+.nut-switch-label-close-disabled {
+  color: #c2c4cc;
 }
 .nut-switch-disabled {
   background-color: #ffadbe;
 }
 .nut-switch-disabled-close {
   background-color: #f5f6fa;
-}
-.nut-switch-label {
-  color: #ffffff;
-  font-size: 12px;
-}
-.nut-switch-label-open {
-  transform: translateX(-18px);
-}
-.nut-switch-label-open-rtl {
-  transform: translateX(18px);
-}
-.nut-switch-label-close {
-  transform: translateX(18px);
-}
-.nut-switch-label-close-rtl {
-  transform: translateX(-18px);
 }
