+++ src/packages/switch/switch.taro.tsx
@@ -10,8 +10,8 @@ export interface SwitchProps extends BasicComponent {
   checked: boolean
   defaultChecked: boolean
   disabled: boolean
-  activeText: string
-  inactiveText: string
+  activeText: React.ReactNode
+  inactiveText: React.ReactNode
   onChange: (
     val: boolean,
     event: React.MouseEvent<Element, MouseEvent> | ITouchEvent
@@ -90,25 +90,20 @@ export const Switch: FunctionComponent<Partial<SwitchProps>> = (props) => {
         {!value && !activeText && (
           <View className={`${classPrefix}-close-line`} />
         )}
-        {activeText && (
-          <View
-            className={classNames([
-              [`${classPrefix}-label`],
-              [
-                value
-                  ? `${classPrefix}-label-open`
-                  : `${classPrefix}-label-close`,
-              ],
-              {
-                [`${classPrefix}-label-open-rtl`]: rtl && value,
-                [`${classPrefix}-button-close-rtl`]: rtl && !value,
-              },
-            ])}
-          >
-            {value ? activeText : inactiveText}
-          </View>
-        )}
       </View>
+      {activeText && (
+        <View
+          className={classNames(`${classPrefix}-label`, {
+            [`${classPrefix}-label-open`]: value,
+            [`${classPrefix}-label-close`]: !value,
+            [`${classPrefix}-label-open-rtl`]: rtl && value,
+            [`${classPrefix}-label-close-rtl`]: rtl && !value,
+            [`${classPrefix}-label-close-disabled`]: disabled && !value,
+          })}
+        >
+          {value ? activeText : inactiveText}
+        </View>
+      )}
     </View>
   )
 }
